# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  active: 'active',
  exported: 'exported',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  totalFrames?: number | null;
  /** @nullable */
  durationMs?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  fps?: number;
  totalFrames?: number;
  durationMs?: number;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  draft: 'draft',
  active: 'active',
  exported: 'exported',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  fps?: number;
  totalFrames?: number;
  durationMs?: number;
}

export interface ProjectStats {
  sourceCount: number;
  blurRegionCount: number;
  frameEditCount: number;
  maskedRegionCount: number;
}

export type VideoSourceRole = typeof VideoSourceRole[keyof typeof VideoSourceRole];


export const VideoSourceRole = {
  primary: 'primary',
  comparison: 'comparison',
  reference: 'reference',
} as const;

export interface VideoSource {
  id: number;
  projectId: number;
  filename: string;
  role: VideoSourceRole;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  totalFrames?: number | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type VideoSourceInputRole = typeof VideoSourceInputRole[keyof typeof VideoSourceInputRole];


export const VideoSourceInputRole = {
  primary: 'primary',
  comparison: 'comparison',
  reference: 'reference',
} as const;

export interface VideoSourceInput {
  filename: string;
  role: VideoSourceInputRole;
  fps?: number;
  totalFrames?: number;
  durationMs?: number;
  notes?: string;
}

export type VideoSourceUpdateRole = typeof VideoSourceUpdateRole[keyof typeof VideoSourceUpdateRole];


export const VideoSourceUpdateRole = {
  primary: 'primary',
  comparison: 'comparison',
  reference: 'reference',
} as const;

export interface VideoSourceUpdate {
  role?: VideoSourceUpdateRole;
  fps?: number;
  totalFrames?: number;
  durationMs?: number;
  notes?: string;
}

export type BlurRegionBlurType = typeof BlurRegionBlurType[keyof typeof BlurRegionBlurType];


export const BlurRegionBlurType = {
  face: 'face',
  body: 'body',
  region: 'region',
  custom: 'custom',
} as const;

export interface BlurRegion {
  id: number;
  projectId: number;
  label: string;
  blurType: BlurRegionBlurType;
  frameStart: number;
  frameEnd: number;
  trackingEnabled: boolean;
  masked: boolean;
  /**
     * JSON-encoded array of per-frame bounding box / polygon coordinates
     * @nullable
     */
  regionData?: string | null;
  createdAt: string;
}

export type BlurRegionInputBlurType = typeof BlurRegionInputBlurType[keyof typeof BlurRegionInputBlurType];


export const BlurRegionInputBlurType = {
  face: 'face',
  body: 'body',
  region: 'region',
  custom: 'custom',
} as const;

export interface BlurRegionInput {
  label: string;
  blurType: BlurRegionInputBlurType;
  frameStart: number;
  frameEnd: number;
  trackingEnabled?: boolean;
  masked?: boolean;
  regionData?: string;
}

export type BlurRegionUpdateBlurType = typeof BlurRegionUpdateBlurType[keyof typeof BlurRegionUpdateBlurType];


export const BlurRegionUpdateBlurType = {
  face: 'face',
  body: 'body',
  region: 'region',
  custom: 'custom',
} as const;

export interface BlurRegionUpdate {
  label?: string;
  blurType?: BlurRegionUpdateBlurType;
  frameStart?: number;
  frameEnd?: number;
  trackingEnabled?: boolean;
  masked?: boolean;
  regionData?: string;
}

export type FrameEditEditType = typeof FrameEditEditType[keyof typeof FrameEditEditType];


export const FrameEditEditType = {
  replace: 'replace',
  merge: 'merge',
  skip: 'skip',
} as const;

export interface FrameEdit {
  id: number;
  projectId: number;
  frameIndex: number;
  /**
     * Source video to pull this frame from (for merge edits)
     * @nullable
     */
  sourceId: number | null;
  editType: FrameEditEditType;
  /**
     * Base64 data URL of replacement image (for replace edits)
     * @nullable
     */
  replacementDataUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type FrameEditInputEditType = typeof FrameEditInputEditType[keyof typeof FrameEditInputEditType];


export const FrameEditInputEditType = {
  replace: 'replace',
  merge: 'merge',
  skip: 'skip',
} as const;

export interface FrameEditInput {
  frameIndex: number;
  sourceId?: number;
  editType: FrameEditInputEditType;
  replacementDataUrl?: string;
  notes?: string;
}