# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Video Editor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Video editing projects
  - name: sources
    description: Video source metadata per project
  - name: blurRegions
    description: Blur / privacy mask regions per project
  - name: frameEdits
    description: Frame replacement/edit records per project
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List video sources for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoSource"
    post:
      operationId: addSource
      tags: [sources]
      summary: Add a video source to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoSourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoSource"

  /projects/{id}/sources/{sourceId}:
    delete:
      operationId: removeSource
      tags: [sources]
      summary: Remove a video source from a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
    patch:
      operationId: updateSource
      tags: [sources]
      summary: Update source metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sourceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoSourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoSource"

  /projects/{id}/blur-regions:
    get:
      operationId: listBlurRegions
      tags: [blurRegions]
      summary: List blur regions for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blur regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlurRegion"
    post:
      operationId: createBlurRegion
      tags: [blurRegions]
      summary: Create a blur region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlurRegionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlurRegion"

  /projects/{id}/blur-regions/{regionId}:
    patch:
      operationId: updateBlurRegion
      tags: [blurRegions]
      summary: Update a blur region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlurRegionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlurRegion"
    delete:
      operationId: deleteBlurRegion
      tags: [blurRegions]
      summary: Delete a blur region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/frame-edits:
    get:
      operationId: listFrameEdits
      tags: [frameEdits]
      summary: List frame edits for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Frame edits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FrameEdit"
    post:
      operationId: createFrameEdit
      tags: [frameEdits]
      summary: Record a frame edit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameEditInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrameEdit"

  /projects/{id}/frame-edits/{editId}:
    delete:
      operationId: deleteFrameEdit
      tags: [frameEdits]
      summary: Remove a frame edit (revert to original)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: editId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get summary stats for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, name, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, exported]
        fps:
          type: ["number", "null"]
        totalFrames:
          type: ["integer", "null"]
        durationMs:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        fps:
          type: number
        totalFrames:
          type: integer
        durationMs:
          type: integer

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [draft, active, exported]
        fps:
          type: number
        totalFrames:
          type: integer
        durationMs:
          type: integer

    ProjectStats:
      type: object
      required: [sourceCount, blurRegionCount, frameEditCount, maskedRegionCount]
      properties:
        sourceCount:
          type: integer
        blurRegionCount:
          type: integer
        frameEditCount:
          type: integer
        maskedRegionCount:
          type: integer

    VideoSource:
      type: object
      required: [id, projectId, filename, role, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        filename:
          type: string
        role:
          type: string
          enum: [primary, comparison, reference]
        fps:
          type: ["number", "null"]
        totalFrames:
          type: ["integer", "null"]
        durationMs:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    VideoSourceInput:
      type: object
      required: [filename, role]
      properties:
        filename:
          type: string
        role:
          type: string
          enum: [primary, comparison, reference]
        fps:
          type: number
        totalFrames:
          type: integer
        durationMs:
          type: integer
        notes:
          type: string

    VideoSourceUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [primary, comparison, reference]
        fps:
          type: number
        totalFrames:
          type: integer
        durationMs:
          type: integer
        notes:
          type: string

    BlurRegion:
      type: object
      required: [id, projectId, label, blurType, frameStart, frameEnd, trackingEnabled, masked, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        label:
          type: string
        blurType:
          type: string
          enum: [face, body, region, custom]
        frameStart:
          type: integer
        frameEnd:
          type: integer
        trackingEnabled:
          type: boolean
        masked:
          type: boolean
        regionData:
          type: ["string", "null"]
          description: JSON-encoded array of per-frame bounding box / polygon coordinates
        createdAt:
          type: string

    BlurRegionInput:
      type: object
      required: [label, blurType, frameStart, frameEnd]
      properties:
        label:
          type: string
        blurType:
          type: string
          enum: [face, body, region, custom]
        frameStart:
          type: integer
        frameEnd:
          type: integer
        trackingEnabled:
          type: boolean
        masked:
          type: boolean
        regionData:
          type: string

    BlurRegionUpdate:
      type: object
      properties:
        label:
          type: string
        blurType:
          type: string
          enum: [face, body, region, custom]
        frameStart:
          type: integer
        frameEnd:
          type: integer
        trackingEnabled:
          type: boolean
        masked:
          type: boolean
        regionData:
          type: string

    FrameEdit:
      type: object
      required: [id, projectId, frameIndex, sourceId, editType, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        frameIndex:
          type: integer
        sourceId:
          type: ["integer", "null"]
          description: Source video to pull this frame from (for merge edits)
        editType:
          type: string
          enum: [replace, merge, skip]
        replacementDataUrl:
          type: ["string", "null"]
          description: Base64 data URL of replacement image (for replace edits)
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    FrameEditInput:
      type: object
      required: [frameIndex, editType]
      properties:
        frameIndex:
          type: integer
        sourceId:
          type: integer
        editType:
          type: string
          enum: [replace, merge, skip]
        replacementDataUrl:
          type: string
        notes:
          type: string