# lib/api-zod/src/generated/types/blurRegion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */
import type { BlurRegionBlurType } from './blurRegionBlurType';

export interface BlurRegion {
  id: number;
  projectId: number;
  label: string;
  blurType: BlurRegionBlurType;
  frameStart: number;
  frameEnd: number;
  trackingEnabled: boolean;
  masked: boolean;
  /**
     * JSON-encoded array of per-frame bounding box / polygon coordinates
     * @nullable
     */
  regionData?: string | null;
  createdAt: string;
}