# lib/api-zod/src/generated/types/blurRegionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */
import type { BlurRegionInputBlurType } from './blurRegionInputBlurType';

export interface BlurRegionInput {
  label: string;
  blurType: BlurRegionInputBlurType;
  frameStart: number;
  frameEnd: number;
  trackingEnabled?: boolean;
  masked?: boolean;
  regionData?: string;
}