# lib/api-zod/src/generated/types/blurRegionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */
import type { BlurRegionUpdateBlurType } from './blurRegionUpdateBlurType';

export interface BlurRegionUpdate {
  label?: string;
  blurType?: BlurRegionUpdateBlurType;
  frameStart?: number;
  frameEnd?: number;
  trackingEnabled?: boolean;
  masked?: boolean;
  regionData?: string;
}