# lib/api-zod/src/generated/types/frameEdit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */
import type { FrameEditEditType } from './frameEditEditType';

export interface FrameEdit {
  id: number;
  projectId: number;
  frameIndex: number;
  /**
     * Source video to pull this frame from (for merge edits)
     * @nullable
     */
  sourceId: number | null;
  editType: FrameEditEditType;
  /**
     * Base64 data URL of replacement image (for replace edits)
     * @nullable
     */
  replacementDataUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}