# lib/api-zod/src/generated/types/frameEditInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */
import type { FrameEditInputEditType } from './frameEditInputEditType';

export interface FrameEditInput {
  frameIndex: number;
  sourceId?: number;
  editType: FrameEditInputEditType;
  replacementDataUrl?: string;
  notes?: string;
}