# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Video Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  sourceCount: number;
  blurRegionCount: number;
  frameEditCount: number;
  maskedRegionCount: number;
}